@@ -19,9 +19,11 @@ import com.google.android.exoplayer2.PlaybackException;
 import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.Player.PositionInfo;
-import com.google.android.exoplayer2.SimpleExoPlayer;
+import com.google.android.exoplayer2.ExoPlayer;
 import com.google.android.exoplayer2.Timeline;
+import com.google.android.exoplayer2.TracksInfo;
 import com.google.android.exoplayer2.audio.AudioAttributes;
+import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory;
 import com.google.android.exoplayer2.metadata.Metadata;
 import com.google.android.exoplayer2.metadata.MetadataOutput;
 import com.google.android.exoplayer2.metadata.icy.IcyHeaders;
@@ -34,12 +36,11 @@ import com.google.android.exoplayer2.source.ShuffleOrder;
 import com.google.android.exoplayer2.source.ShuffleOrder.DefaultShuffleOrder;
 import com.google.android.exoplayer2.source.SilenceMediaSource;
 import com.google.android.exoplayer2.source.TrackGroup;
-import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.source.dash.DashMediaSource;
 import com.google.android.exoplayer2.source.hls.HlsMediaSource;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.upstream.DataSource;
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
+import com.google.android.exoplayer2.upstream.DefaultDataSource;
 import com.google.android.exoplayer2.upstream.DefaultHttpDataSource;
 import com.google.android.exoplayer2.util.MimeTypes;
 import com.google.android.exoplayer2.util.Util;
@@ -96,7 +97,8 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
     private int lastPlaylistLength = 0;
     private Map<String, Object> pendingPlaybackEvent;
 
-    private SimpleExoPlayer player;
+    private ExoPlayer player;
+    private DefaultExtractorsFactory extractorsFactory = new DefaultExtractorsFactory();
     private Integer audioSessionId;
     private MediaSource mediaSource;
     private Integer currentIndex;
@@ -138,6 +140,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
         eventChannel = new BetterEventChannel(messenger, "com.ryanheise.just_audio.events." + id);
         dataEventChannel = new BetterEventChannel(messenger, "com.ryanheise.just_audio.data." + id);
         processingState = ProcessingState.none;
+        extractorsFactory.setConstantBitrateSeekingEnabled(true);
         if (audioLoadConfiguration != null) {
             Map<?, ?> loadControlMap = (Map<?, ?>)audioLoadConfiguration.get("androidLoadControl");
             if (loadControlMap != null) {
@@ -214,9 +217,9 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
     }
 
     @Override
-    public void onTracksChanged(TrackGroupArray trackGroups, TrackSelectionArray trackSelections) {
-        for (int i = 0; i < trackGroups.length; i++) {
-            TrackGroup trackGroup = trackGroups.get(i);
+    public void onTracksInfoChanged(TracksInfo tracks) {
+        for (int i = 0; i < tracks.getTrackGroupInfos().size(); i++) {
+            TrackGroup trackGroup = tracks.getTrackGroupInfos().get(i).getTrackGroup();
 
             for (int j = 0; j < trackGroup.length; j++) {
                 Metadata metadata = trackGroup.getFormat(j).metadata;
@@ -272,14 +275,14 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
         if (player.getPlaybackState() == Player.STATE_ENDED) {
             try {
                 if (player.getPlayWhenReady()) {
-                    if (player.hasNextWindow()) {
-                        player.seekToNextWindow();
+                    if (player.hasNextMediaItem()) {
+                        player.seekToNextMediaItem();
                     } else if (lastPlaylistLength == 0 && player.getMediaItemCount() > 0) {
                         player.seekTo(0, 0L);
                     }
                 } else {
-                    if (player.getCurrentWindowIndex() < player.getMediaItemCount()) {
-                        player.seekTo(player.getCurrentWindowIndex(), 0L);
+                    if (player.getCurrentMediaItemIndex() < player.getMediaItemCount()) {
+                        player.seekTo(player.getCurrentMediaItemIndex(), 0L);
                     }
                 }
             } catch (Exception e) {
@@ -290,7 +293,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
     }
 
     private boolean updateCurrentIndex() {
-        Integer newIndex = player.getCurrentWindowIndex();
+        Integer newIndex = player.getCurrentMediaItemIndex();
         // newIndex is never null.
         // currentIndex is sometimes null.
         if (!newIndex.equals(currentIndex)) {
@@ -380,7 +383,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
             sendError(String.valueOf(error.errorCode), error.getMessage());
         }
         errorCount++;
-        if (player.hasNextWindow() && currentIndex != null && errorCount <= 5) {
+        if (player.hasNextMediaItem() && currentIndex != null && errorCount <= 5) {
             int nextIndex = currentIndex + 1;
             Timeline timeline = player.getCurrentTimeline();
             // This condition is due to: https://github.com/ryanheise/just_audio/pull/310
@@ -587,7 +590,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
         String id = (String)map.get("id");
         switch ((String)map.get("type")) {
         case "progressive":
-            return new ProgressiveMediaSource.Factory(buildDataSourceFactory())
+            return new ProgressiveMediaSource.Factory(buildDataSourceFactory(), extractorsFactory)
                     .createMediaSource(new MediaItem.Builder()
                             .setUri(Uri.parse((String)map.get("uri")))
                             .setTag(id)
@@ -686,7 +689,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
         DataSource.Factory httpDataSourceFactory = new DefaultHttpDataSource.Factory()
             .setUserAgent(userAgent)
             .setAllowCrossProtocolRedirects(true);
-        return new DefaultDataSourceFactory(context, httpDataSourceFactory);
+        return new DefaultDataSource.Factory(context, httpDataSourceFactory);
     }
 
     private void load(final MediaSource mediaSource, final long initialPosition, final Integer initialIndex, final Result result) {
@@ -717,7 +720,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
 
     private void ensurePlayerInitialized() {
         if (player == null) {
-            SimpleExoPlayer.Builder builder = new SimpleExoPlayer.Builder(context);
+            ExoPlayer.Builder builder = new ExoPlayer.Builder(context);
             if (loadControl != null) {
                 builder.setLoadControl(loadControl);
             }
@@ -955,7 +958,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
         seekPos = position;
         seekResult = result;
         try {
-            int windowIndex = index != null ? index : player.getCurrentWindowIndex();
+            int windowIndex = index != null ? index : player.getCurrentMediaItemIndex();
             player.seekTo(windowIndex, position);
         } catch (RuntimeException e) {
             seekResult = null;
@@ -1006,7 +1009,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
     // depending on the number of bits required to
     // represent it.
     public static Long getLong(Object o) {
-        return (o == null || o instanceof Long) ? (Long)o : new Long(((Integer)o).intValue());
+        return (o == null || o instanceof Long) ? (Long)o : Long.valueOf(((Integer)o).intValue());
     }
 
     @SuppressWarnings("unchecked")
