@@ -19,10 +19,10 @@ class SeekBar extends StatefulWidget {
   }) : super(key: key);
 
   @override
-  _SeekBarState createState() => _SeekBarState();
+  SeekBarState createState() => SeekBarState();
 }
 
-class _SeekBarState extends State<SeekBar> {
+class SeekBarState extends State<SeekBar> {
   double? _dragValue;
   late SliderThemeData _sliderThemeData;
 
@@ -180,3 +180,5 @@ void showSliderDialog({
     ),
   );
 }
+
+T? ambiguate<T>(T? value) => value;
